Encryption finished

+++ src/galois.c
@@ -1,23 +1,23 @@
 #include "galois.h"
 
-unsigned int multiplication(unsigned a, unsigned int b);
-unsigned int redux(unsigned int a, unsigned int b);
+uint16_t multiplication(uint8_t a, uint8_t b);
+uint8_t redux(uint16_t a, uint16_t b);
 
-unsigned int sum(unsigned int a, unsigned int b){
+uint8_t sum(uint8_t a, uint8_t b){
     return a ^ b;
 }
 
 
-unsigned int prod(unsigned int a, unsigned int b){
+uint8_t prod(uint8_t a, uint8_t b){
 
-    unsigned int result = multiplication(a,b);
+    uint16_t result = multiplication(a,b);
 
     return redux(result, POL_PRIMITIVE);
 
 
 }
 
-unsigned int multiplication(unsigned a, unsigned int b){
+uint16_t multiplication(uint8_t a, uint8_t b){
     
     int i = 0;
     int array[8];
@@ -25,38 +25,38 @@ unsigned int multiplication(unsigned a, unsigned int b){
         array[i] = (a >> i) & 1;
     }
     
-    unsigned int result = 0;
+    uint16_t result = 0;
 
     for (i = 0; i < PIXEL_SIZE; i++){
         if(array[i] == 1){
-            result = result ^ (b << i);
+            result = result ^ ((uint16_t) b << i);
         }
     }
     
     return result;
 }
 
-unsigned int recus_degree(unsigned int a, int exp){
+int recus_degree(uint16_t a, int exp){
     if ( a / pow(2, exp + 1) < 1){
         return exp;
     }
     return recus_degree(a, exp + 1);
 }
 
-unsigned int degree(unsigned int a){
+int degree(uint16_t a){
     if (a == 0){
         return 0;
     }
     return recus_degree(a, 0);
 }
 
-unsigned int redux(unsigned int p, unsigned int m){
+uint8_t redux(uint16_t p, uint16_t m){
     int deg_p = degree(p);
     int deg_m = degree(m);
     int deg_resto;
 
     int deg = deg_p - deg_m;
-    unsigned int resto = p;
+    uint16_t resto = p;
 
     while(deg >= 0){
         resto = (m << deg) ^ resto;
@@ -65,16 +65,16 @@ unsigned int redux(unsigned int p, unsigned int m){
     
         deg = deg_resto - deg_m;
     }
-
-     return resto;
+    
+     return (uint8_t) resto;
 }
 
-unsigned int gf_pow(unsigned int pixel, int times){
+uint8_t gf_pow(uint8_t pixel, int times){
     if(times == 0){
         return 1;
     }
     int i; 
-    unsigned int result = pixel;
+    uint8_t result = pixel;
 
     for(i = 0; i < times - 1; i++){
         result = prod(result, pixel);
@@ -83,13 +83,13 @@ unsigned int gf_pow(unsigned int pixel, int times){
     return result;
 }
 
-unsigned int f_block(unsigned int * block, unsigned int pixel, int k){
+uint8_t f_block(uint8_t * block, uint8_t pixel, int k){
     int i;
-    unsigned int result = 0;
+    uint8_t result = 0;
 
     for(i = 0; i < k; i++){
-        unsigned int s = block[i];
-        unsigned int aux = prod(s, gf_pow(pixel, i));
+        uint8_t s = block[i];
+        uint8_t aux = prod(s, gf_pow(pixel, i));
         result = sum(result, aux);
     }
